% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PythonPkgWrapperUtils.R
\name{autoGenerateRdFiles}
\alias{autoGenerateRdFiles}
\title{Create the autogenerated files}
\usage{
autoGenerateRdFiles(srcRootDir, functionInfo, classInfo, templateDir = NULL)
}
\arguments{
\item{srcRootDir}{is the root directory for the code base (i.e., prior to installation)}

\item{functionInfo}{list of function to generate docs}

\item{classInfo}{list of class to generate docs}

\item{templateDir}{(optional) custom templates for the docs}
}
\description{
This function generate R docs from python docs and saves the doc files to the target dir <srcRootDir>/auto-man
}
