% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PySource.R
\name{BEGIN.Python}
\alias{BEGIN.Python}
\title{Execute Python interactively from within R}
\usage{
BEGIN.Python()
}
\value{
Returns the entered code back to R, code lines which throw an
        exception are obmitted.
}
\description{
The function BEGIN.Python starts an Python Read–eval–print loop.
}
\details{
BEGIN.Python emulates the behavior of the Python terminal
         and therefore allows interactive Python code development
         from within R.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
\dontrun{
code <-
BEGIN.Python()
import os
os.getcwd()
dir(os)
x = 3**3
END.Python
## NOTE: BEGIN.Python returns the successfully executed code as string.
cat(code)
pyGet0("x")
}
}

