% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyObject.R
\docType{data}
\name{callFun}
\alias{callFun}
\title{}
\format{\preformatted{ chr "\\nfunction(...)\{\\n  x <- list(...)\\n  i <- if ( !is.null(names(x)) ) (nchar(names(x)) > 0) else rep(FALSE, length(x))\\n  xargs "| __truncated__
}}
\usage{
callFun
}
\description{

}
\keyword{datasets}

