% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PythonPkgWrapperUtils.R
\name{cleanUpStackTrace}
\alias{cleanUpStackTrace}
\title{Clean up stack trace}
\usage{
cleanUpStackTrace(callable, args)
}
\arguments{
\item{callable}{the callable to be called}

\item{args}{the input to be passed to callable}
}
\value{
the result of calling callable with args after cleaning up the stack trace
}
\description{
This function clean up stack trace for a callable.
}
