% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PythonPkgWrapperUtils.R
\name{determineArgsAndKwArgs}
\alias{determineArgsAndKwArgs}
\title{Determines args and kwargs}
\usage{
determineArgsAndKwArgs(...)
}
\arguments{
\item{...}{the input values}
}
\value{
A list of args and kwargs
}
\description{
This function determines args and kwargs for a given list of values.
}
