% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pyAttach.R
\name{pyAttach}
\alias{pyAttach}
\title{attach Python objects to R}
\usage{
pyAttach(what, env = parent.frame())
}
\arguments{
\item{what}{a character vector the names which should be attached to R.}

\item{env}{the environment where he virtual Python objects are
           assigned to.}
}
\description{
A convenience function to attach Python objects to R.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
}

