% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyObjectDict.R
\name{pyDict}
\alias{pyDict}
\title{create a virtual Python dictionary}
\usage{
pyDict(key, value, regFinalizer = TRUE)
}
\arguments{
\item{key}{a character string giving the name of the Python object.}

\item{value}{an optional value.}

\item{regFinalizer}{a logical}
}
\description{
The function pyDict creates a virtual Python object
             of type PythonInR_Dict.
}
\details{
TODO
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyExec('myPyDict = {"a":1, "b":2, "c":3}')
# create a virtual Python dictionary for an existing dictionary
myDict <- pyDict("myPyDict")
myDict[""]
myDict[1] <- "should give an error since tuple are not mutable"
myDict
# create a new Python list and virtual list
newTuple <- pyTuple('myNewTuple', list(1:3, 'Hello Python'))
newTuple[1]
}

