% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PyObject.R
\name{pyFunction}
\alias{pyFunction}
\title{create a virtual Python dictionary}
\usage{
pyFunction(key)
}
\arguments{
\item{key}{a string specifying the name of a Python method/function.}
}
\description{
The function pyDict creates a virtual Python object
             of type PythonInR_Dict.
}
\details{
TODO
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
}

