% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyGet.R
\name{pyGet0}
\alias{pyGet0}
\title{creates an R representation of an Python object}
\usage{
pyGet0(key)
}
\arguments{
\item{key}{a string specifying the name of a Python object.}
}
\value{
Returns a virtual python object.
}
\description{
The function pyGet0 gets Python objects by name.
}
\details{
Primitive data types like bool, int, long, float, str,
         bytes and unicode are
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyExec("import os")
os <- pyGet0("os")
os$getcwd()
os$sep
os$sep <- "Hello Python!"
pyExecp("os.sep")
}

