% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Basics.R
\name{pyImport}
\alias{pyImport}
\title{is a convenience function to call the Python function \bold{import}}
\usage{
pyImport(module = character(), alias = "")
}
\arguments{
\item{module}{A string specifying the name of the Python module.}

\item{alias}{an optional string defining a alias for the module name.}
}
\description{
A convenience function to call the Python function \bold{import}.
}
\details{
The function pyImport has a special behavior for the packages numpy
         and pandas. For these two packages pyImport does not only import
         numpy but also register their alias in the options. To be found when
         pyGet and pySet is used with the option useNumpy set to TRUE.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyImport("os")
\dontrun{
#NOTE: The following does not only import numpy but also register the
#      alias in the options under the name "numpyAlias".
#      The same is done for pandas, the default alias for pandas and numpy
#      are respectively "pandas" and "numpy". The numpyAlias is used
#      when calling pyGet or pySet with the option useNumpy set to TRUE.
pyImport("numpy", "np")
pyImport("pandas", "pd")
}
}

