% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PyObject.R
\name{pyObject}
\alias{pyObject}
\title{create a virtual Python dictionary}
\usage{
pyObject(key, regFinalizer = TRUE)
}
\arguments{
\item{key}{a character string giving the name of the Python object.}

\item{regFinalizer}{a logical}
}
\description{
The function pyDict creates a virtual Python object
             of type PythonInR_Dict.
}
\details{
TODO
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
}

