% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PyOptions.R
\name{pyOptions}
\alias{pyOptions}
\title{a options object}
\usage{
pyOptions(option, value)
}
\arguments{
\item{option}{TODO}

\item{value}{TODO}
}
\description{
Set options.
}
\details{
Hash options.
}

