% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PySet.R
\name{pySet}
\alias{pySet}
\title{assign R objects to Python}
\usage{
pySet(key, value, namespace = "__main__", useNumpy = pyOptions("useNumpy"),
  usePandas = pyOptions("usePandas"))
}
\arguments{
\item{key}{a string specifying the name of the Python object.}

\item{value}{an R object which is assigned to Python.}

\item{namespace}{a string specifying where the key should be located.
If the namespace is set to "__main__" the key will be
set to the global namespace. But it is also posible to
set attributes of objects e.g. the attribute name of
the object 'os'.}

\item{useNumpy}{an optional logical, default is FALSE, to control if numpy
should be used for the type conversion of matrices.}

\item{usePandas}{an optional logical, default is FALSE, to control if pandas
should be used for the type conversion of data frames.}
}
\description{
The function pySet allows to assign R objects to the Python
             namespace, the conversion from R to Python is done automatically.
}
\details{
More information about the type conversion can be found in the README
         file or at \url{http://pythoninr.bitbucket.org/}.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pySet("x", 3)
pySet("M", diag(1,3))
pyImport("os")
pySet("name", "Hello os!", namespace="os")
}

