% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PySource.R
\name{pySource}
\alias{pySource}
\title{}
\usage{
pySource(file, local = FALSE, echo = verbose, print.eval = echo,
  verbose = getOption("verbose"), prompt.echo = getOption("prompt"),
  max.deparse.length = 150, chdir = FALSE,
  encoding = getOption("encoding"), continue.echo = getOption("continue"),
  skip.echo = 0, keep.source = getOption("keep.source"))
}
\description{

}

