% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyObjectTuple.R
\name{pyTuple}
\alias{pyTuple}
\title{create a virtual Python tuple}
\usage{
pyTuple(key, value, regFinalizer = FALSE)
}
\arguments{
\item{key}{}

\item{value}{}

\item{regFinalizer}{}
}
\description{
The function pyTuple
}
\details{
TODO
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyExec('myPyTuple = (1, 2, 5, "Hello R!")')
# create a virtual Python tuple for an existing tuple
myTuple <- pyTuple("myPyTuple")
myTuple[0]
myTuple[1] <- "should give an error since tuple are not mutable"
myTuple
# create a new Python list and virtual list
newTuple <- pyTuple('myNewTuple', list(1:3, 'Hello Python'))
newTuple[1]
}

