% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PyObjectDict.R
\name{pyDict}
\alias{pyDict}
\title{Create a virtual Python dictionary}
\usage{
pyDict(key, value, regFinalizer = TRUE)
}
\arguments{
\item{key}{a character string giving the name of the Python object.}

\item{value}{an optional value, allowed values are named vectors,
named lists and NULL.}

\item{regFinalizer}{a logical indicating if a finalizer should be
be registered, the default value is TRUE.}
}
\description{
The function pyDict creates a virtual Python object
             of type PythonInR_Dict.
}
\details{
If no value is provided a virtual Python dict for an existing
         Python object is created. If the value is NULL an empty
         virtual Python object for an empty dict is created.
         If the value is a named vector or named list a new Python
         object based on the vector or list is created.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyExec('myPyDict = {"a":1, "b":2, "c":3}')
# create a virtual Python dictionary for an existing dictionary
myDict <- pyDict("myPyDict")
myDict["a"]
myDict["a"] <- "set the key"
myDict
# create a new Python dict and virtual dict
myNewDict <- pyDict('myNewDict', list(p=2, y=9, r=1))
myNewDict
}

