% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyObjectList.R
\name{pyList}
\alias{pyList}
\title{create a virtual Python list}
\usage{
pyList(key, value, regFinalizer = TRUE)
}
\arguments{
\item{key}{}

\item{value}{}

\item{regFinalizer}{}
}
\description{
The function pyList
}
\details{
TODO
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyExec('myPyList = [1, 2, 5, "Hello R!"]')
# create a virtual Python list for an existing list
myList <- pyList("myPyList")
myList[0]
myList[1] <- "changed"
myList
# create a new Python list and virtual list
pyList('myNewList', list(1:3, 'Hello Python'))
pyList[1]
}

