% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SetPythonObjects.R
\name{pySet}
\alias{pySet}
\title{assign R objects to Python}
\usage{
pySet(key, value, useNumpy = getOption("useNumpy"),
  usePandas = getOption("usePandas"))
}
\arguments{
\item{key}{a string specifying the name of the Python object.}

\item{value}{an R object which is assigned to Python.}

\item{useNumpy}{an optional logical, default is FALSE, to control if numpy
should be used for the type conversion of matrices.}

\item{usePandas}{an optional logical, default is FALSE, to control if pandas
should be used for the type conversion of data frames.}
}
\description{
The function pySet allows to assign R objects to the Python
             namespace, the conversion from R to Python is done automatically.
}
\details{
More information about the type conversion can be found in the README
         file or at \url{http://pythoninr.bitbucket.org/}.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pySet("x", 3)
pySet("M", diag(1,3))
}

